refactor: FormContainer -> FormBox로 네이밍 변경
- 기본의 Container 컴포넌트와 혼동이 올 것 같아 단순 컴포넌트인
FormContainer는 FormBox로 이름 변경
