+++ frontend/.eslintrc.js
@@ -21,5 +21,6 @@ module.exports = {
   rules: {
     'react/jsx-filename-extension': 0,
     'react/jsx-tag-spacing': 0,
+    'import/prefer-default-export': 0,
   },
 };
